social_media_api/requirements.txt
# Django core
Django>=6.0,<7.0

# Django REST Framework
djangorestframework>=3.14

# Token Authentication
djangorestframework-authtoken>=1.3

# Database
psycopg2-binary>=2.9  # PostgreSQL adapter
dj-database-url>=1.0

# Deployment
gunicorn>=20.1
whitenoise>=6.5

# Optional: environment variable handling
python-decouple>=3.8

# Misc
pillow>=9.5  # For handling profile images
